+++ Assets/RoDOTS/Runtime/Controllers/GameController.cs
@@ -1,0 +1,8 @@
+﻿using EugeneC.Singleton;
+
+namespace RoDOTS.runtime {
+    
+    public class GameController : GenericSingleton<GameController> {
+        
+    }
+}