+++ Packages/com.datea83.choyutilities/Runtime/Entities/Agents/AgentMoveSystemAuthoring.cs
@@ -6,6 +6,7 @@ namespace EugeneC.ECS {
     [DisallowMultipleComponent]
     public class AgentMoveSystemAuthoring : MonoBehaviour {
         
+        [SerializeField][Min(0.01f)] private float rotateSpeed = 1f;
         [SerializeField][Min(0.01f)] private float minSpeed = 1f, maxSpeed = 10f;
         [SerializeField] private bool hasRestTime = true;
         [SerializeField][Min(0.01f)] private float minRestTime = 1f, maxRestTime = 5f;
@@ -23,12 +24,15 @@ namespace EugeneC.ECS {
             public override void Bake(AgentMoveSystemAuthoring authoring) {
                 var e = GetEntity(TransformUsageFlags.None);
                 
+                AddComponent<AgentSpawnISingleton>(e);
+                
                 AddComponent(e, new AgentMoveSystemISingleton {
                     MinSpeed = authoring.minSpeed,
                     MaxSpeed = authoring.maxSpeed,
                     MinRestTime = authoring.minRestTime,
                     MaxRestTime = authoring.maxRestTime,
-                    HasRestTime = authoring.hasRestTime
+                    HasRestTime = authoring.hasRestTime,
+                    RotationSpeed = authoring.rotateSpeed
                 });
             }
         }