+++ Packages/com.datea83.choyutilities/Runtime/Entities/Agents/AgentSpawnNodeAuthoring.cs
@@ -43,9 +43,14 @@ namespace EugeneC.ECS {
     [UpdateBefore(typeof(InitializeAgentMoveISystem))]
     public partial struct AgentSpawnISystem : ISystem {
 
+        public void OnCreate(ref SystemState state) {
+            state.RequireForUpdate<AgentSpawnISingleton>();
+        }
+
         [BurstCompile]
         public void OnUpdate(ref SystemState state) {
             var ecb = new EntityCommandBuffer(state.WorldUpdateAllocator);
+            var singleton = SystemAPI.GetSingleton<AgentSpawnISingleton>();
             var et = SystemAPI.Time.ElapsedTime;
             var dt = SystemAPI.Time.DeltaTime;
             
@@ -59,6 +64,8 @@ namespace EugeneC.ECS {
 
                 var index = entity.RandomValue(et, 0, buffer.Length);
                 var newEntity = ecb.Instantiate(buffer[index].Prefab);
+                singleton.CurrentSpawnCount++;
+                singleton.TotalSpawnCount++;
                 
                 ecb.AddComponent(newEntity, new InitializeAgentIData {
                     Spawn = entity,
@@ -71,6 +78,8 @@ namespace EugeneC.ECS {
                 }
             }
             ecb.Playback(state.EntityManager);
+            
+            SystemAPI.SetSingleton(singleton);
         }
 
     }