+++ Packages/com.datea83.choyutilities/Runtime/Entities/UI/UiHandleSystemBase.cs
@@ -7,9 +7,9 @@ namespace EugeneC.ECS {
     [UpdateInGroup(typeof(Eu_EffectSystemGroup))]
     public partial class UiHandleSystemBase : SystemBase {
 
-        public event Action<byte, byte> OnUiEnterHover;
-        public event Action<byte, byte> OnUiExitHover;
-        public event Action<byte, byte> OnUiClicked;
+        // public event Action<byte, byte> OnUiEnterHover;
+        // public event Action<byte, byte> OnUiExitHover;
+        // public event Action<byte, byte> OnUiClicked;
 
         protected override void OnCreate() { }